src/feature-extractor/pyproject.toml
[tool.poetry]
name = "feature-extractor"
version = "0.1.0"
description = "Set of tools to extract features from specs and files."
authors = ["Dimitri Podborski"]

[tool.poetry.dependencies]
python = "^3.9"
beautifulsoup4 = "^4.10.0"
lxml = "^4.6.3"
GitPython = "^3.1.24"
openpyxl = "^3.0.9"
xmltodict = "^0.12.0"
python-docx = "^0.8.11"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
extract-spec-features = "feature_extractor.spec_features:extract_spec_features"
update-spec-features = "feature_extractor.spec_features:update_spec_features"
check-spec-features = "feature_extractor.spec_features:check_spec_features"

contribute-files = "feature_extractor.file_features:contribute_files"
extract-file-boxes-gpac = "feature_extractor.file_features:extract_file_boxes_gpac"
update-heif-features = "feature_extractor.file_features:update_heif_features"
update-ff-conformance-xls = "feature_extractor.file_features:update_ff_conformance_xls"

extract-html-features = "feature_extractor.html_features:extract_html_features"